refactor: Moved welcome and check updates to another file
